+++ core/src/commonMain/kotlin/com/powersync/utils/Log.kt
@@ -1,24 +1,28 @@
 package com.powersync.utils
 
+import BuildConfig
 import co.touchlab.kermit.Logger
 import co.touchlab.kermit.Severity
+import co.touchlab.kermit.StaticConfig
+import co.touchlab.kermit.platformLogWriter
 
 /*
  * Generates a logger with the appropriate severity level based on the build type
  * if no Logger is provided.
 */
-public fun generateLogger(logger: Logger?): Logger {
-    if (logger != null) {
-        return logger
-    }
-
-    val defaultLogger: Logger = Logger
-
-    if (BuildConfig.isDebug) {
-        Logger.setMinSeverity(Severity.Verbose)
-    } else {
-        Logger.setMinSeverity(Severity.Warn)
-    }
-
-    return defaultLogger
-}
+public fun generateLogger(logger: Logger?): Logger =
+    logger
+        ?: Logger(
+            config =
+                StaticConfig(
+                    logWriterList =
+                        listOf(platformLogWriter()),
+                    minSeverity =
+                        if (BuildConfig.isDebug) {
+                            Severity.Verbose
+                        } else {
+                            Severity.Warn
+                        },
+                ),
+            tag = "PowerSync",
+        )