+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Add `PowerSyncDatabase.getCrudTransactions()`, returning a flow of transactions. This is useful
   to upload multiple transactions in a batch.
+* Fix modifying severity of the global Kermit logger
+* Add `PowerSync` tag for the logs
 
 ## 1.4.0